MapCreator UserControl Updates
ConfigureColorTables: added a button so users can load the Terrain.xml. This is useful so users can edit, or add to, the games terrain TileID numbers and associate them with the Adobe Photoshop™ color table and swatches.

Added and linked up a credits screen so that credit is given where credit was earned. Accessible by pressing the blue information button on the top right of the main form after launch.